9043/refactor: extracted user related methods to the UserService
