+++ core/src/main/java/greencity/controller/ManagementEmployeeController.java
@@ -13,8 +13,8 @@ import greencity.dto.position.PositionDto;
 import greencity.dto.tariff.GetTariffInfoForEmployeeDto;
 import greencity.filters.EmployeeFilterCriteria;
 import greencity.filters.EmployeePage;
-import greencity.service.ubs.UBSClientService;
 import greencity.service.ubs.UBSManagementEmployeeService;
+import greencity.service.ubs.user.UserService;
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.Parameter;
 import io.swagger.v3.oas.annotations.media.ArraySchema;
@@ -25,16 +25,25 @@ import io.swagger.v3.oas.annotations.responses.ApiResponses;
 import jakarta.validation.Valid;
 import jakarta.validation.constraints.Email;
 import jakarta.validation.constraints.Positive;
+import java.util.List;
+import java.util.Set;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.MediaType;
 import org.springframework.http.ResponseEntity;
 import org.springframework.security.access.prepost.PreAuthorize;
 import org.springframework.validation.annotation.Validated;
-import org.springframework.web.bind.annotation.*;
+import org.springframework.web.bind.annotation.DeleteMapping;
+import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.PathVariable;
+import org.springframework.web.bind.annotation.PostMapping;
+import org.springframework.web.bind.annotation.PutMapping;
+import org.springframework.web.bind.annotation.RequestBody;
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RequestParam;
+import org.springframework.web.bind.annotation.RequestPart;
+import org.springframework.web.bind.annotation.RestController;
 import org.springframework.web.multipart.MultipartFile;
-import java.util.List;
-import java.util.Set;
 
 @RestController
 @RequestMapping("/admin/ubs-employee")
@@ -42,7 +51,7 @@ import java.util.Set;
 @Validated
 public class ManagementEmployeeController {
     private final UBSManagementEmployeeService employeeService;
-    private final UBSClientService ubsClientService;
+    private final UserService userService;
 
     /**
      * Saves a new employee with optional image upload.
@@ -219,7 +228,7 @@ public class ManagementEmployeeController {
     @GetMapping("/get-all-authorities")
     public ResponseEntity<Object> getAllAuthorities(
         @Email(regexp = ValidationConstant.EMAIL_REGEXP) @RequestParam String email) {
-        Set<String> authorities = ubsClientService.getAllAuthorities(email);
+        Set<String> authorities = userService.getAllAuthorities(email);
         return ResponseEntity.status(HttpStatus.OK).body(authorities);
     }
 
@@ -244,7 +253,7 @@ public class ManagementEmployeeController {
     @GetMapping("/get-positions-authorities")
     public ResponseEntity<PositionAuthoritiesDto> getPositionsAndRelatedAuthorities(
         @Email(regexp = ValidationConstant.EMAIL_REGEXP) @RequestParam String email) {
-        return ResponseEntity.status(HttpStatus.OK).body(ubsClientService.getPositionsAndRelatedAuthorities(email));
+        return ResponseEntity.status(HttpStatus.OK).body(userService.getPositionsAndRelatedAuthorities(email));
     }
 
     /**
@@ -265,7 +274,7 @@ public class ManagementEmployeeController {
     @PreAuthorize("@preAuthorizer.hasAuthority('EDIT_EMPLOYEES_AUTHORITIES', authentication)")
     @PutMapping("/edit-authorities")
     public ResponseEntity<Object> editAuthorities(@Valid @RequestBody UserEmployeeAuthorityDto dto) {
-        ubsClientService.updateEmployeesAuthorities(dto);
+        userService.updateEmployeesAuthorities(dto);
         return ResponseEntity.status(HttpStatus.OK).build();
     }
