+++ core/src/main/java/greencity/controller/UserProfileController.java
@@ -6,7 +6,7 @@ import greencity.dto.user.DeactivateUserRequestDto;
 import greencity.dto.user.UserProfileCreateDto;
 import greencity.dto.user.UserProfileDto;
 import greencity.dto.user.UserProfileUpdateDto;
-import greencity.service.ubs.UBSClientService;
+import greencity.service.ubs.user.UserService;
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.Parameter;
 import io.swagger.v3.oas.annotations.media.Content;
@@ -18,14 +18,19 @@ import lombok.RequiredArgsConstructor;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.validation.annotation.Validated;
-import org.springframework.web.bind.annotation.*;
+import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.PostMapping;
+import org.springframework.web.bind.annotation.PutMapping;
+import org.springframework.web.bind.annotation.RequestBody;
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RestController;
 
 @RestController
 @Validated
 @RequestMapping("/ubs/userProfile")
 @RequiredArgsConstructor
 public class UserProfileController {
-    private final UBSClientService ubsClientService;
+    private final UserService userService;
 
     /**
      * Controller returns user`s data or update {@link UserProfileDto} date.
@@ -48,7 +53,7 @@ public class UserProfileController {
         @Parameter(hidden = true) @CurrentUserUuid String userUuid,
         @Valid @RequestBody UserProfileUpdateDto userProfileUpdateDto) {
         return ResponseEntity.status(HttpStatus.OK)
-            .body(ubsClientService.updateProfileData(userUuid, userProfileUpdateDto));
+            .body(userService.updateProfileData(userUuid, userProfileUpdateDto));
     }
 
     /**
@@ -67,7 +72,7 @@ public class UserProfileController {
     @GetMapping("/user/getUserProfile")
     public ResponseEntity<UserProfileDto> getUserData(
         @Parameter(hidden = true) @CurrentUserUuid String userUuid) {
-        return ResponseEntity.status(HttpStatus.OK).body(ubsClientService.getProfileData(userUuid));
+        return ResponseEntity.status(HttpStatus.OK).body(userService.getProfileData(userUuid));
     }
 
     /**
@@ -88,7 +93,7 @@ public class UserProfileController {
     @PostMapping("/user/create")
     public ResponseEntity<Long> createUserProfile(
         @Valid @RequestBody UserProfileCreateDto userProfileCreateDto) {
-        return ResponseEntity.status(HttpStatus.CREATED).body(ubsClientService.createUserProfile(userProfileCreateDto));
+        return ResponseEntity.status(HttpStatus.CREATED).body(userService.createUserProfile(userProfileCreateDto));
     }
 
     /**
@@ -108,7 +113,7 @@ public class UserProfileController {
     public ResponseEntity<HttpStatus> deactivateUser(
         @Parameter(hidden = true) @CurrentUserUuid String uuid,
         @Valid @RequestBody DeactivateUserRequestDto request) {
-        ubsClientService.markUserAsDeactivated(uuid, request);
+        userService.markUserAsDeactivated(uuid, request);
         return ResponseEntity.status(HttpStatus.OK).build();
     }
 }
